refactor: moved test files to cutr package
